+++ sdl-core/pom.xml
@@ -356,7 +356,7 @@
 		<dependency>
 			<groupId>org.glassfish.jersey.media</groupId>
 			<artifactId>jersey-media-json-jackson</artifactId>
-			<version>2.34</version>
+			<version>${jersey.version}</version>
 		</dependency>
 
 		<!--Jetty Websocket server side dependencies start here -->
